+++ tests/testthat/test-datatransfer.R
@@ -16,7 +16,7 @@ test_that("Testing data transfer between R and athena", {
   
   df <- data.frame(w = as.POSIXct((Sys.time() -9):Sys.time(), origin = "1970-01-01"),
                    x = 1:10,
-                   y = letters[1:10], 
+                   y = c(letters[1:8], c(" \\t\\t\\n 123 \" \\t\\t\\n ", ",15 \"")), 
                    z = sample(c(TRUE, FALSE), 10, replace = T),
                    stringsAsFactors = F)
   
@@ -41,8 +41,8 @@ test_that("Testing data transfer between R and athena", {
   test_df2 <- as.data.frame(dbGetQuery(con, paste0("select w, x, y, z from test_df2 where year = '", format(DATE, "%Y"), "' and month = '",format(DATE, "%m"), "' and day = '", format(DATE, "%d"),"'")))
   test_df3 <- as.data.frame(dbGetQuery(con, "select * from df_bigint"))
   test_df4 <- as.data.frame(dbGetQuery(con, "select * from mtcars2"))
-  expect_equal(test_df,df)
-  expect_equal(test_df2,df)
+  expect_equal(test_df,sqlData(con, df))
+  expect_equal(test_df2,sqlData(con, df))
   expect_equal(test_df3,df2)
   expect_equal(test_df4, sqlData(con, mtcars))